requirements.txt
Django==2.1.3
django-filter==2.0.0
djangorestframework==3.9.0
pytz==2018.7
